The old conv network is in action_net some work will need to be done to make it match the current eval code as it was made for working with the first iteration of my eval last fall
